+++ HabboHotel/Rooms/RoomManager.cs
@@ -30,6 +30,7 @@ public class RoomManager : IRoomManager
     private readonly IItemLoader _itemLoader;
     private readonly IRoomItemPersistenceService _roomItemPersistenceService;
     private readonly IRoomItemPlacementValidatorService _roomItemPlacementValidatorService;
+    private readonly IRoomItemPlacementPersistenceService _roomItemPlacementPersistenceService;
     private readonly IChatManager _chatManager;
     private readonly IBotManager _botManager;
     private readonly IRoomService _roomService;
@@ -53,6 +54,7 @@ public class RoomManager : IRoomManager
         IItemLoader itemLoader,
         IRoomItemPersistenceService roomItemPersistenceService,
         IRoomItemPlacementValidatorService roomItemPlacementValidatorService,
+        IRoomItemPlacementPersistenceService roomItemPlacementPersistenceService,
         IGameClientManager gameClientManager,
         IGroupManager groupManager,
         IRoomService roomService,
@@ -74,6 +76,7 @@ public class RoomManager : IRoomManager
         _itemLoader = itemLoader;
         _roomItemPersistenceService = roomItemPersistenceService;
         _roomItemPlacementValidatorService = roomItemPlacementValidatorService;
+        _roomItemPlacementPersistenceService = roomItemPlacementPersistenceService;
         _clientManager = gameClientManager;
         _groupManager = groupManager;
         _roomService = roomService;
@@ -215,7 +218,7 @@ public class RoomManager : IRoomManager
 
     private Room CreateRoomInstance(RoomData data)
     {
-        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
+        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
     }
 
     private static void DisposeRoom(Room room) => room.Dispose();